Created all of the basic entities. - added builders - main mappers - moved repeating code to ContextUtil - updated views

+++ src/main/java/ua/denitdao/servlet/shop/model/dao/mapper/EntityMapper.java
@@ -8,5 +8,5 @@ public interface EntityMapper<T> {
 
     T extractFromResultSet(ResultSet resultSet) throws SQLException;
 
-    T makeUnique(Map<Integer, T> cache, T entity);
+    T makeUnique(Map<Long, T> cache, T entity);
 }
